# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Workout Tracker BE API
  version: "1.0.0"
  summary: Offline-first sync API for the mobile workout tracker.
  description: |
    Bearer-token API for syncing workout progress and plans between the devices
    of one user, plus a shared exercise catalogue.

    Conventions
    -----------
    - Base path `/v1`; all bodies are `application/json; charset=utf-8` unless
      stated otherwise.
    - Authentication is an opaque bearer token (`Authorization: Bearer <token>`),
      valid for 365 days. Only `POST /v1/auth/login` and `GET /healthz` are
      anonymous.
    - Two roles: `user` and `admin`. Admin accounts are management-only (they
      cannot use workout data endpoints) and are limited to endpoints 1, 13, 14
      plus login/logout for their own session.
    - IDs are lowercase UUID strings. Progress and workout plan IDs are
      client-generated; exercises accept an optional client-generated ID.
    - Timestamps are RFC 3339 UTC (`2026-09-19T08:30:00Z`).
    - Field names are `snake_case`. Unknown request fields are rejected (400).
    - Decimal weights (`kg`, max 2 places) and distances (`m`, max 2 places) are
      JSON numbers, never strings. `rpe` is a number in 0.5 steps from 1 to 10.
    - For client-written rows, `updated_at` is the client's last-modified time
      (the conflict-rule input) and `server_updated_at` is the server sync
      cursor. `deleted_at` is nullable (soft delete).
...
      and is the only non-JSON body.
...

...
servers:
...
paths:
...
  /v1/progress:
    get:
      operationId: listProgress
      tags: [progress]
      summary: List the caller's completed sessions
      description: |
        Doubles as the sync pull endpoint via `updated_since`. Summaries by
        default; `expand=exercises` returns full sessions (with exercises and
        sets). With `updated_since`, rows are ordered by `server_updated_at`
        ascending and soft-deleted rows are included.

        Auth: bearer token, role `user` (admin gets 403).
      security:
        - bearerAuth: []
      x-required-role: user
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/UpdatedSince'
        - $ref: '#/components/parameters/WorkoutPlanIDFilter'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/IncludeDeleted'
        - $ref: '#/components/parameters/Expand'
      responses:
        '200':
          description: |
            A page of summaries, or of full sessions when `expand=exercises`.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ProgressSummaryPage'
                  - $ref: '#/components/schemas/ProgressPage'
              examples:
                syncExpanded:
                  summary: Sync pull with updated_since and expand=exercises
                  value:
                    items:
                      - id: 0195f3a2-cccc-7000-8000-000000000100
                        workout_plan_id: 0195f3a2-aaaa-7000-8000-000000000001
                        name: Push Day A
                        notes: Felt strong
                        started_at: '2026-09-19T08:00:00Z'
                        ended_at: '2026-09-19T08:55:00Z'
                        duration_seconds: 3120
                        updated_at: '2026-09-19T08:55:02Z'
                        created_at: '2026-09-19T08:55:04Z'
                        server_updated_at: '2026-09-19T08:55:04Z'
                        deleted_at: null
                        exercises:
                          - exercise_id: 0195f3a2-bbbb-7000-8000-000000000010
                            position: 0
                            notes: null
                            sets:
                              - position: 0
                                type: normal
                                reps: 8
                                weight: 80.0
                                duration_seconds: null
                                distance_meters: null
                                rpe: 8.5
                                completed: true
                    next_cursor: null
                summary:
                  summary: Default summary page
                  value:
                    items:
                      - id: 0195f3a2-cccc-7000-8000-000000000100
                        workout_plan_id: 0195f3a2-aaaa-7000-8000-000000000001
                        name: Push Day A
                        started_at: '2026-09-19T08:00:00Z'
                        ended_at: '2026-09-19T08:55:00Z'
                        duration_seconds: 3120
                        exercise_count: 5
                        set_count: 18
                        updated_at: '2026-09-19T08:55:02Z'
                        server_updated_at: '2026-09-19T08:55:04Z'
                        deleted_at: null
                    next_cursor: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/progress/{id}:
    parameters:
      - $ref: '#/components/parameters/PathID'
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get one completed session
      description: |
        Full session with exercises and sets. A missing, soft-deleted or
        foreign-owned session is 404.

        Auth: bearer token, role `user` (admin gets 403).
      security:
        - bearerAuth: []
      x-required-role: user
      responses:
        '200':
          description: The session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progress'
              examples:
                session:
                  value:
                    id: 0195f3a2-cccc-7000-8000-000000000100
                    workout_plan_id: 0195f3a2-aaaa-7000-8000-000000000001
                    name: Push Day A
                    notes: Felt strong
                    started_at: '2026-09-19T08:00:00Z'
                    ended_at: '2026-09-19T08:55:00Z'
                    duration_seconds: 3120
                    updated_at: '2026-09-19T08:55:02Z'
                    created_at: '2026-09-19T08:55:04Z'
                    server_updated_at: '2026-09-19T08:55:04Z'
                    deleted_at: null
                    exercises:
                      - exercise_id: 0195f3a2-bbbb-7000-8000-000000000010
                        position: 0
                        notes: null
                        sets:
                          - position: 0
                            type: normal
                            reps: 8
                            weight: 80.0
                            duration_seconds: null
                            distance_meters: null
                            rpe: 8.5
                            completed: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
    put:
      operationId: saveProgress
      tags: [progress]
      summary: Create or update a completed session
      description: |
        Idempotent upsert keyed by the client-generated `id`. The request is the
        full state of the session (replace, not patch).

        Conflict rule: no row -> 201; row soft-deleted -> 409 `deleted`; request
        `updated_at` older than stored -> 409 `stale` with the server copy in
        `error.details[0].current`; equal -> 200 no-op returning the stored copy;
        newer -> 200 saved. `updated_at` more than 5 minutes in the future is 422.

        Auth: bearer token, role `user` (admin gets 403).
      security:
        - bearerAuth: []
      x-required-role: user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressInput'
            examples:
              saveProgress:
                summary: Save a completed session
                value:
                  workout_plan_id: 0195f3a2-aaaa-7000-8000-000000000001
                  name: Push Day A
                  notes: Felt strong
                  started_at: '2026-09-19T08:00:00Z'
                  ended_at: '2026-09-19T08:55:00Z'
                  duration_seconds: 3120
                  updated_at: '2026-09-19T08:55:02Z'
                  exercises:
                    - exercise_id: 0195f3a2-bbbb-7000-8000-000000000010
                      position: 0
                      notes: null
                      sets:
                        - position: 0
                          type: warmup
                          reps: 12
                          weight: 40.0
                          duration_seconds: null
                          distance_meters: null
                          rpe: null
                          completed: true
                        - position: 1
                          type: normal
                          reps: 8
                          weight: 80.0
                          duration_seconds: null
                          distance_meters: null
                          rpe: 8.5
                          completed: true
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progress'
        '200':
          description: Session updated, or no-op when `updated_at` equals the stored value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Progress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/Internal'
    delete:
      operationId: deleteProgress
      tags: [progress]
      summary: Soft-delete a completed session
      description: |
        Sets `deleted_at` and `server_updated_at`; children are kept. Idempotent:
        deleting an already-deleted session also returns 204. A later `PUT` to
        the same id is 409 `deleted`.

        Auth: bearer token, role `user` (admin gets 403).
      security:
        - bearerAuth: []
      x-required-role: user
      responses:
        '204':
          description: Session deleted (or already deleted).
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/Internal'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque token issued by `POST /v1/auth/login`. Send it as
        `Authorization: Bearer <access_token>`. Expired, revoked or unknown
        tokens get 401 with `WWW-Authenticate: Bearer`.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size, 1..200. Defaults to 50; out of range is 422.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque `next_cursor` from the previous page. A malformed cursor is 400.
      schema:
        type: string
    UpdatedSince:
      name: updated_since
      in: query
      required: false
      description: |
        RFC 3339 timestamp, or a `yyyy-MM-dd` date meaning midnight UTC of that
        day. Returns rows whose sync cursor is strictly after it, including
        soft-deleted rows, ordered for sync. Any other value is 400. Sync
        clients should send the exact `server_updated_at` they last saw, so a
        pull does not repeat the whole day. Note: a `+` offset must be
        URL-encoded as `%2B`.
      schema:
        type: string
        description: RFC 3339 timestamp or `yyyy-MM-dd` date.
        example: "2026-09-19T08:30:00Z"
    IncludeDeleted:
      name: include_deleted
      in: query
      required: false
      description: |
        Include soft-deleted rows in the default order. Defaults to false;
        `updated_since` implies true. Only the literal `true`/`false` are
        accepted (else 422).
      schema:
        type: boolean
        default: false
    Expand:
      name: expand
      in: query
      required: false
      description: |
        `exercises` returns full objects (progress sessions or plans with their
        exercises) instead of summaries. Any other value is 422.
      schema:
        type: string
        enum: [exercises]
...
    WorkoutPlanIDFilter:
      name: workout_plan_id
      in: query
      required: false
      description: Filter progress sessions by workout plan. A non-UUID value is 400.
      schema:
        type: string
        format: uuid
    From:
      name: from
      in: query
      required: false
      description: |
        Inclusive lower bound on `started_at`. RFC 3339 timestamp, or a
        `yyyy-MM-dd` date meaning midnight UTC of that day. To filter by the
        device's local day, send local midnight with its offset (URL-encode `+`
        as `%2B`).
      schema:
        type: string
        description: RFC 3339 timestamp or `yyyy-MM-dd` date.
        example: "2026-09-19"
    To:
      name: to
      in: query
      required: false
      description: |
        Inclusive upper bound on `started_at`. RFC 3339 timestamp, or a
        `yyyy-MM-dd` date meaning the end of that UTC day, so `to=2026-09-19`
        includes the whole day.
      schema:
        type: string
        description: RFC 3339 timestamp or `yyyy-MM-dd` date.
        example: "2026-09-19"
    PathID:
      name: id
      in: path
      required: true
      description: Resource UUID (canonical hyphenated form). A malformed value is 400.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 'Malformed request (400 `bad_request`): bad JSON, unknown field, bad path id, bad cursor or timestamp.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            badRequest:
              value:
                error:
                  code: bad_request
                  message: The request is malformed.
    Unauthorized:
      description: 'Missing, invalid, expired or revoked token, or bad credentials (401 `unauthorized`).'
      headers:
        WWW-Authenticate:
          description: Always `Bearer`.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error:
                  code: unauthorized
                  message: Authentication is required.
    Forbidden:
      description: 'Authenticated but not allowed (403 `forbidden`), e.g. wrong role.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            forbidden:
              value:
                error:
                  code: forbidden
                  message: You are not allowed to perform this action.
    NotFound:
      description: 'Missing resource, or a resource owned by another user (404 `not_found`; existence is not leaked).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              value:
                error:
                  code: not_found
                  message: The resource was not found.
    Conflict:
      description: |
        Conflict (409 `conflict`): duplicate (`already_taken`, `already_exists`,
        `id_taken`), stale write (`stale`, with the server copy in
        `error.details[0].current`) or write to a deleted row (`deleted`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            stale:
              summary: Stale progress write; the server copy is returned
              value:
                error:
                  code: conflict
                  message: The request conflicts with the current state of the resource.
                  details:
                    - issue: stale
                      current:
                        id: 0195f3a2-cccc-7000-8000-000000000100
                        workout_plan_id: 0195f3a2-aaaa-7000-8000-000000000001
                        name: Push Day A
                        notes: Felt strong
                        started_at: '2026-09-19T08:00:00Z'
                        ended_at: '2026-09-19T08:55:00Z'
                        duration_seconds: 3120
                        updated_at: '2026-09-19T08:55:02Z'
                        created_at: '2026-09-19T08:55:04Z'
                        server_updated_at: '2026-09-19T08:55:04Z'
                        deleted_at: null
                        exercises: []
            deleted:
              summary: Write to a soft-deleted row
              value:
                error:
                  code: conflict
                  message: The request conflicts with the current state of the resource.
                  details:
                    - issue: deleted
    PayloadTooLarge:
      description: 'Body over the limit (413 `payload_too_large`): 1 MiB for a JSON body, 2 MiB for the image upload.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            tooLarge:
              value:
                error:
                  code: payload_too_large
                  message: The request body is too large.
    UnsupportedMediaType:
      description: 'Wrong `Content-Type` (415 `unsupported_media_type`): JSON routes require `application/json`; the image upload requires a supported image type that matches the sniffed bytes.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unsupported:
              value:
                error:
                  code: unsupported_media_type
                  message: The content type is not supported.
    ValidationFailed:
      description: 'Well-formed input with invalid values (422 `validation_failed`); `details` lists every problem.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              value:
                error:
                  code: validation_failed
                  message: The request has invalid values.
                  details:
                    - field: username
                      issue: invalid_format
...
    Internal:
      description: 'Unexpected server error (500 `internal`). No internals are leaked.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            internal:
              value:
                error:
                  code: internal
                  message: Internal server error.
  schemas:
...
    ErrorCode:
      type: string
      description: Stable machine-readable error code.
      enum:
        - bad_request
        - unauthorized
        - forbidden
        - not_found
        - conflict
        - payload_too_large
        - unsupported_media_type
        - validation_failed
        - rate_limited
        - internal
    ErrorDetail:
      type: object
      required: [issue]
      properties:
        field:
          type: string
          description: JSON field path (may be empty or absent for whole-request issues).
        issue:
          type: string
          description: Stable machine issue value, e.g. `stale`, `deleted`, `already_exists`, `required`, `invalid_value`.
        existing_id:
          type: string
          format: uuid
          description: Other exercise's id for an `already_exists` conflict.
        current:
          description: |
            Server's current copy for a `stale` conflict, in the response shape
            of the resource (a `Progress` or a `Plan`).
          oneOf:
            - $ref: '#/components/schemas/Progress'
            - $ref: '#/components/schemas/Plan'
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Human-readable summary; may change between releases.
            details:
              type: array
              description: Optional list of problems or conflict metadata.
              items:
                $ref: '#/components/schemas/ErrorDetail'

    # --- auth ---------------------------------------------------------------
...
    SetType:
      type: string
      enum: [warmup, normal, drop, failure]

    # --- decimals -----------------------------------------------------------
    Weight:
      type: number
      description: Weight in kg, 0 to 99999.99, at most 2 decimals.
      minimum: 0
      maximum: 99999.99
      multipleOf: 0.01
    DistanceMeters:
      type: number
      description: Distance in meters, 0 to 9999999.99, at most 2 decimals.
      minimum: 0
      maximum: 9999999.99
      multipleOf: 0.01
    Rpe:
      type: number
      description: Rate of perceived exertion, 1 to 10 in 0.5 steps.
      minimum: 1
      maximum: 10
      multipleOf: 0.5

    # --- exercises ----------------------------------------------------------
...
    PlanExercise:
      type: object
      required: [exercise_id, position, target_sets]
      properties:
        exercise_id:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0
        target_sets:
          type: integer
          minimum: 1
          maximum: 20
        target_reps:
          type: [integer, "null"]
          minimum: 1
        target_reps_max:
          type: [integer, "null"]
          minimum: 1
        target_weight:
          oneOf:
            - $ref: '#/components/schemas/Weight'
            - type: "null"
        target_duration_seconds:
          type: [integer, "null"]
          minimum: 0
        target_distance_meters:
          oneOf:
            - $ref: '#/components/schemas/DistanceMeters'
            - type: "null"
        rest_seconds:
          type: [integer, "null"]
          minimum: 0
          maximum: 3600
        notes:
          type: [string, "null"]
    Plan:
      type: object
      required:
        - id
        - name
        - updated_at
        - created_at
        - server_updated_at
        - exercises
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        updated_at:
          type: string
          format: date-time
          description: Client's last-modified time (conflict-rule input).
        created_at:
          type: string
          format: date-time
        server_updated_at:
          type: string
          format: date-time
          description: Server sync cursor; set on every accepted write and delete.
        deleted_at:
          type: [string, "null"]
          format: date-time
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/PlanExercise'
...
    ProgressSetInput:
      type: object
      required: [position, type, completed]
      additionalProperties: false
      properties:
        position:
          type: integer
          minimum: 0
        type:
          $ref: '#/components/schemas/SetType'
        reps:
          type: [integer, "null"]
          minimum: 0
        weight:
          oneOf:
            - $ref: '#/components/schemas/Weight'
            - type: "null"
        duration_seconds:
          type: [integer, "null"]
          minimum: 0
        distance_meters:
          oneOf:
            - $ref: '#/components/schemas/DistanceMeters'
            - type: "null"
        rpe:
          oneOf:
            - $ref: '#/components/schemas/Rpe'
            - type: "null"
        completed:
          type: boolean
    ProgressExerciseInput:
      type: object
      required: [exercise_id, position, sets]
      additionalProperties: false
      properties:
        exercise_id:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0
        notes:
          type: [string, "null"]
          maxLength: 1000
        sets:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ProgressSetInput'
    ProgressInput:
      type: object
      required: [name, started_at, ended_at, duration_seconds, updated_at, exercises]
      additionalProperties: false
      properties:
        workout_plan_id:
          type: [string, "null"]
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 100
        notes:
          type: [string, "null"]
          maxLength: 2000
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          description: Must be >= `started_at`.
        duration_seconds:
          type: integer
          minimum: 0
          maximum: 86400
        updated_at:
          type: string
          format: date-time
          description: Client's last-modified time; conflict-rule input. At most 5 minutes in the future.
        exercises:
          type: array
          maxItems: 50
          description: May be empty. Positions must be unique within their parent.
          items:
            $ref: '#/components/schemas/ProgressExerciseInput'
    ProgressSet:
      type: object
      required: [position, type, completed]
      properties:
        position:
          type: integer
          minimum: 0
        type:
          $ref: '#/components/schemas/SetType'
        reps:
          type: [integer, "null"]
          minimum: 0
        weight:
          oneOf:
            - $ref: '#/components/schemas/Weight'
            - type: "null"
        duration_seconds:
          type: [integer, "null"]
          minimum: 0
        distance_meters:
          oneOf:
            - $ref: '#/components/schemas/DistanceMeters'
            - type: "null"
        rpe:
          oneOf:
            - $ref: '#/components/schemas/Rpe'
            - type: "null"
        completed:
          type: boolean
    ProgressExercise:
      type: object
      required: [exercise_id, position, sets]
      properties:
        exercise_id:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 0
        notes:
          type: [string, "null"]
        sets:
          type: array
          items:
            $ref: '#/components/schemas/ProgressSet'
    Progress:
      type: object
      required:
        - id
        - name
        - started_at
        - ended_at
        - duration_seconds
        - updated_at
        - created_at
        - server_updated_at
        - exercises
      properties:
        id:
          type: string
          format: uuid
        workout_plan_id:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        notes:
          type: [string, "null"]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
          minimum: 0
          maximum: 86400
        updated_at:
          type: string
          format: date-time
          description: Client's last-modified time (conflict-rule input).
        created_at:
          type: string
          format: date-time
        server_updated_at:
          type: string
          format: date-time
          description: Server sync cursor; set on every accepted write and delete.
        deleted_at:
          type: [string, "null"]
          format: date-time
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ProgressExercise'
    ProgressSummary:
      type: object
      required:
        - id
        - name
        - started_at
        - ended_at
        - duration_seconds
        - exercise_count
        - set_count
        - updated_at
        - server_updated_at
      properties:
        id:
          type: string
          format: uuid
        workout_plan_id:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        duration_seconds:
          type: integer
          minimum: 0
        exercise_count:
          type: integer
          minimum: 0
        set_count:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
        server_updated_at:
          type: string
          format: date-time
        deleted_at:
          type: [string, "null"]
          format: date-time
    ProgressPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Progress'
        next_cursor:
          type: [string, "null"]
    ProgressSummaryPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProgressSummary'
        next_cursor:
          type: [string, "null"]

    # --- ops ----------------------------------------------------------------
...